requirements.txt
﻿# --- Core Django & Config ---
Django==5.1.4
python-decouple==3.8        # Substitui o decouple para alinhar com o settings
gunicorn==23.0.0        # Servidor Web
google-generativeai     # AI Vision (Scanner de Encartes)
# --- API Rest (Nomes Corrigidos) ---
djangorestframework
djangorestframework-simplejwt
django-cors-headers
# --- Banco de Dados (PostgreSQL) ---
psycopg2-binary         # O driver correto para o PythonJet
dj-database-url==3.0.1  # Conexão automática
# --- Estáticos e Mídia ---
pillow==11.3.0          # Para imagens
# --- Autenticação ---
django-allauth==65.10.0
python3-openid==3.2.0
requests-oauthlib==2.0.0
oauthlib==3.3.1
PyJWT==2.10.1
cryptography==45.0.5
# --- Pagamentos ---
mercadopago==2.3.0
requests==2.32.4
urllib3==2.5.0
certifi==2025.7.14
idna==3.10
charset-normalizer==3.4.2
# --- Tarefas (Celery) ---
celery==5.5.3
kombu==5.5.4
amqp==5.3.1
billiard==4.2.1
vine==5.1.0
# --- PDF (Relatórios) ---
WeasyPrint
reportlab
# --- Dependências Gerais ---
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
packaging==25.0
Faker==37.5.3
python-dateutil==2.9.0.post0
six==1.17.0
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
wcwidth==0.2.13
prompt_toolkit==3.0.51
cffi==1.17.1
whitenoise